requirements.txt
torch
torchvision
numpy
opencv-python
pillow
streamlit
requests
